Add files via upload
We generated 3 prior dataset, and find the best parameter of BDC, plot figures of true signals, p_filter and BDC(best). Further more, we make detail compare using dataset3(compare prior of beta, cluster, radius, lambda and iterations. ) And find that, all the parameters should not be too big or too small(we find nearly best as showed in figures of best1.2.3), And find that, when cluster radius of signals are small, the parameters are more sensitive. Best parameters will be slightly different.